requirements.txt
arabic-reshaper==2.1.3
asn1crypto==1.5.1
autopep8==1.6.0
beautifulsoup4==4.10.0
bs4==0.0.1
build==0.7.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cryptography==37.0.1
cssselect2==0.6.0
future==0.18.2
html5lib==1.1
idna==3.3
Jinja2==3.1.2
lxml==4.8.0
MarkupSafe==2.1.1
oscrypto==1.3.0
packaging==21.3
pep517==0.12.0
Pillow==9.1.0
pycodestyle==2.8.0
pycparser==2.21
pyHanko==0.13.0
pyhanko-certvalidator==0.19.5
pyparsing==3.0.8
PyPDF3==1.0.6
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.9
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
svglib==1.2.1
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tqdm==4.64.0
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.9
webencodings==0.5.1
xhtml2pdf==0.2.7
